fix(indexer): fix indexer keys assert
