--- tests/jinad/integration/distributed/conftest.py
@@ -1,15 +1,0 @@
-import os
-import time
-import pytest
-
-
-@pytest.fixture()
-def docker_compose(request):
-    os.system(
-        f"docker-compose -f {request.param} --project-directory . up  --build -d --remove-orphans"
-    )
-    time.sleep(10)
-    yield
-    os.system(
-        f"docker-compose -f {request.param} --project-directory . down --remove-orphans"
-    )
